+++ tests/testthat/test_checkFormat.R
@@ -33,7 +33,7 @@ test_that("Rounding condition", {
   p2 <- checkFormat(p, type="other")
 
   expect_equal(p1$dat$ma_pv1,round(p1$dat$ma_pv1,16))
-  expect_message(p2 <- checkFormat(p1), "Variable ma_pv1 has more decimals than SPSS allows (18) and will be rounded to 16 decimal places.", fixed=TRUE)
+  expect_message(p3 <- checkFormat(p), "Variable ma_pv1 has more decimals than SPSS allows (18) and will be rounded to 16 decimal places.", fixed=TRUE)
   expect_equal(p2$dat$ma_pv1,round(p2$dat$ma_pv1,18))
 
   # No Character Variables
@@ -44,4 +44,26 @@ test_that("Rounding condition", {
 p <- extractVars(pisa, c("idstud"))
 test_that("No Change", {
   expect_message(p3 <- checkFormat(p, changeFormat = FALSE), "Format mismatch for Variable idstud: F8.0 vs. F3", fixed=TRUE)
+})
+
+
+test_that("Empty variables", {
+  # preparation
+  suppressMessages(g3 <- g2 <- g1 <- checkFormat(g))
+
+  g1$dat$var1 <- NA_integer_
+  expect_message(out1 <- checkFormat(g1), "Format of Variable var1 will be changed from A4 to A3", fixed=TRUE)
+  expect_equal(out1$labels$format, c("A3", "A3", "F2", "F2", "F2", "F4.2", "F4.2", "F16.14", "A146"))
+  expect_true(is.character(out1$dat$var1))
+
+  g2$dat$var1 <- NA
+  expect_message(out2 <- checkFormat(g2), "Format of Variable var1 will be changed from A4 to A3", fixed=TRUE)
+  expect_equal(out2$labels$format, c("A3", "A3", "F2", "F2", "F2", "F4.2", "F4.2", "F16.14", "A146"))
+  expect_true(is.character(out2$dat$var1))
+
+  g3$dat$var1 <- NA
+  g3 <- removeValLabels(g3, varName = "var1", value = c(-99, -96))
+  expect_silent(out3 <- checkFormat(g3))
+  expect_equal(out3$labels$format, c("A4", "F2", "F2", "F2", "F4.2", "F4.2", "F16.14", "A146"))
+  expect_true(is.character(out3$dat$var1))
 })